+++ examples/basic-mpa/src/pages/About/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './index.css';
 
 const About = () => {
   return (